requirements.txt
Flask
pdfplumber
python-docx
python-dotenv
google-generativeai
